requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-repeat
allure-pytest
selenium
pyyaml
